refactor: rename stories
