+++ src/components/display/collapsingActions/CollapsingActions.mdx
@@ -1,10 +1,10 @@
 import { Meta, Title, Controls, Stories, Primary, Canvas } from '@storybook/addon-docs/blocks';
 import * as CollapsingActionsStories from './CollapsingActions.stories';
-import CollapsingActionsAlignmentStorySrc from './CollapsingActions.alignment?raw';
-import CollapsingActionsMaxVisibleStorySrc from './CollapsingActions.maxVisible?raw';
-import CollapsingActionsSizeStorySrc from './CollapsingActions.size?raw';
-import CollapsingActionsColorStorySrc from './CollapsingActions.color?raw';
-import CollapsingActionsGapStorySrc from './CollapsingActions.gap?raw';
+import CollapsingActionsAlignmentStorySrc from './CollapsingActions.stories.alignment?raw';
+import CollapsingActionsMaxVisibleStorySrc from './CollapsingActions.stories.maxVisible?raw';
+import CollapsingActionsSizeStorySrc from './CollapsingActions.stories.size?raw';
+import CollapsingActionsColorStorySrc from './CollapsingActions.stories.color?raw';
+import CollapsingActionsGapStorySrc from './CollapsingActions.stories.gap?raw';
 
 <Meta of={CollapsingActionsStories} />
