+++ src/components/display/collapsingActions/CollapsingActions.stories.maxVisible.tsx
