+++ src/components/display/popover/Popover.mdx
@@ -1,7 +1,7 @@
 import { Canvas,Meta, Controls, Stories, Title, Primary } from '@storybook/addon-docs/blocks';
 import * as PopoverStories from './Popover.stories';
-import PopoverOnClickStorySrc from './Popover.onClick?raw';
-import PopoverOnHoverStorySrc from './Popover.onHover?raw';
+import PopoverOnClickStorySrc from './Popover.stories.onClick?raw';
+import PopoverOnHoverStorySrc from './Popover.stories.onHover?raw';
 
 <Meta of={PopoverStories} />
