+++ src/components/display/popper/Popper.stories.onHover.tsx
@@ -1,8 +1,14 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Zextras <https://www.zextras.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
 import React, { useRef, useState } from 'react';
+
+import { Popper } from './Popper';
 import { Button } from '../../basic/button/Button';
 import { Text } from '../../basic/text/Text';
 import { Container } from '../../layout/container/Container';
-import { Popper } from './Popper';
 
 export const PopperOnHover = (): React.JSX.Element => {
 	const [open, setOpen] = useState(false);
@@ -22,7 +28,7 @@ export const PopperOnHover = (): React.JSX.Element => {
 				anchorEl={buttonRef}
 				placement="right"
 				onClose={() => setOpen(false)}
-				disableRestoreFocus={true}
+				disableRestoreFocus
 			>
 				<Text>This is the content of the Popper</Text>
 			</Popper>
